# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Yuppers API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/codes": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sends a one-time code to an email address or phone number. Answers the\nsame way whether or not an account exists for it. Codes sent earlier keep\nworking until they expire, up to the newest few.",
        "operationId": "request_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "A code was sent"
          },
          "422": {
            "description": "Not an email address or phone number (`INVALID_IDENTIFIER`), or a phone number of a country the service does not take (`PHONE_COUNTRY_NOT_SERVED`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many codes requested for this identifier or from this address (`TOO_MANY_REQUESTS`), or too many wrong codes for this identifier today, so none is sent (`TOO_MANY_GUESSES`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/session": {
      "delete": {
        "tags": [
          "auth"
        ],
        "summary": "Signs out: the session stops working everywhere it was held.",
        "operationId": "delete_session",
        "responses": {
          "204": {
            "description": "Signed out"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/sessions": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Signs in with a one-time code, creating the account on first use.",
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSession"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionCreated"
                }
              }
            }
          },
          "401": {
            "description": "The code is wrong, expired or used up",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "The account is suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many wrong codes for this identifier today (`TOO_MANY_GUESSES`), or a wrong code from an address that has offered too many this hour (`TOO_MANY_REQUESTS`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/blocks": {
      "get": {
        "tags": [
          "safety"
        ],
        "summary": "The people the caller has blocked, most recently blocked first.",
        "operationId": "blocked_people",
        "responses": {
          "200": {
            "description": "Each by an exchange shared with them",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockedPerson"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/exchanges/{id}/block": {
      "get": {
        "tags": [
          "safety"
        ],
        "summary": "Whether the caller has blocked the other party of this exchange.",
        "operationId": "block_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exchange ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The caller's own block, if any",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockStatus"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such exchange for this account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "safety"
        ],
        "summary": "Blocks the other party of this exchange. Neither of the two can then join\nan exchange the other starts. Whatever is waiting to be signed between\nthem is withdrawn or declined; an agreement in force stays in force. The\nother party is not told.",
        "operationId": "block",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exchange ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Blocked"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such exchange for this account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Nobody has joined the exchange yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "safety"
        ],
        "summary": "Removes the caller's block on the other party of this exchange.",
        "operationId": "unblock",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exchange ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Not blocked any more"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such exchange for this account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Nobody has joined the exchange yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/exchanges/{id}/commands": {
      "post": {
        "tags": [
          "exchanges"
        ],
        "summary": "Does one thing to an exchange: accept, decline or withdraw a revision,\nact on a contribution, confirm or reject whoever claimed the invitation,\ndiscard a draft, or end or close it.",
        "operationId": "run_command",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exchange ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Unique per attempt to make this change",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Done; the exchange as it now stands",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeView"
                }
              }
            }
          },
          "403": {
            "description": "That is for the other party to do",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such exchange for this account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Refused by the rules, or the exchange changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/exchanges/{id}/draft": {
      "put": {
        "tags": [
          "exchanges"
        ],
        "summary": "Saves the caller's unsent working copy.",
        "operationId": "save_draft",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exchange ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Saved"
          },
          "404": {
            "description": "No such exchange for this account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The exchange is closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/exchanges/{id}/history": {
      "get": {
        "tags": [
          "record"
        ],
        "summary": "What has happened in an exchange, oldest first, with what the parties\nwrote along the way: the message sent with a revision, a note on a claim,\nthe reason for a dispute, a statement. Answers with the latest events, 50\nunless `limit` says otherwise and never more than 200; `earlier` in the\nanswer is the `before` that reads the page before.",
        "operationId": "history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exchange ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Return events before this sequence number. Leave out for the latest.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many events to return: 50 unless said, 200 at most.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of the history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPage"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such exchange for this account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The query is not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/exchanges/{id}/invitation": {
      "post": {
        "tags": [
          "exchanges"
        ],
        "summary": "Replaces the invitation link. The previous link stops working.",
        "operationId": "reissue_invitation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exchange ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationOptions"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The new link token, shown once",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationIssued"
                }
              }
            }
          },
          "403": {
            "description": "Only the initiator can invite",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Someone is in the invited party's place",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/exchanges/{id}/reports": {
      "post": {
        "tags": [
          "safety"
        ],
        "summary": "Reports an exchange, and with it the other party, for review. The other\nparty is not told. Reporting the same exchange again while the first\nreport is open changes nothing and is answered the same way.",
        "operationId": "report_exchange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exchange ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Received"
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such exchange for this account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Nobody has joined the exchange yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Details too long, or missing for `OTHER`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many reports today",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/exchanges/{id}/wallet/apple": {
      "post": {
        "tags": [
          "wallet"
        ],
        "summary": "Adds the caller's pass to Apple Wallet: the signed `.pkpass`. A party to\nthe exchange only, and only once something is agreed.",
        "operationId": "apple_pass",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Exchange ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The pass",
            "content": {
              "application/vnd.apple.pkpass": {
                "schema": {
                  "$ref": "#/components/schemas/Pkpass"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such exchange for this account, or Apple Wallet passes are not available (WALLET_UNAVAILABLE)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Nothing has been agreed yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Asked for too often",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlockStatus": {
        "type": "object",
        "description": "Whether the caller has blocked the other party of an exchange. Says\nnothing about whether the other party has blocked the caller.",
        "required": [
          "blocked",
          "name"
        ],
        "properties": {
          "blocked": {
            "type": "boolean"
          },
          "name": {
            "type": "string",
            "description": "The other party's name as the exchange writes it. An exchange closed\nbefore anything was agreed shows no terms to read it from."
          }
        }
      },
...
      "CreateSession": {
        "type": "object",
        "required": [
          "identifier",
          "code",
          "delivery"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "The one-time code sent to the identifier."
          },
          "delivery": {
            "$ref": "#/components/schemas/Delivery"
          },
          "identifier": {
            "type": "string"
          },
          "language": {
            "type": [
              "string",
              "null"
            ],
            "description": "The language the client is showing, as a tag such as `es-MX`. Used only\nwhen this creates the account; an unsupported one becomes the default."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          }
        }
      },
...
      "RequestCode": {
        "type": "object",
        "required": [
          "identifier"
        ],
        "properties": {
          "identifier": {
            "type": "string",
            "description": "An email address, or a phone number in international form."
          }
        }
      },
...
      "SessionCreated": {
        "type": "object",
        "required": [
          "account"
        ],
        "properties": {
          "account": {
            "$ref": "#/components/schemas/Account"
          },
          "token": {
            "type": [
              "string",
              "null"
            ],
            "description": "Present for `TOKEN` delivery only."
          }
        }
      },
...
    }
  }
}